Add files via upload
The df_analises contain the data which will be manipulated to become a structure, in a table organized with each product ID, value, receipt, supplier and date of arrival (day, month, year). The actual format of the archive is messy and to manipulate it the way it is will be a challenge, if using only Excel and Libre Calc.