Add files via upload
Hey!
The main goal of this project was to manipulate the data of the 'xlsx' archives into a structured table, using Python.
Those reports are generated in an ERP, as you can see, to manipulate the data inside of the files 'analises_v2.xlsx' and 'type_dada.xlsx'. I created the fuction in the Python file 'funcoes_analise.py'. I managed to extract the most important informations: product ID, description, date, suppliers, receipts and values. 

The result of those extractions are in the file 'data_structured.xlsx'.